# specs/ft/football-data/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: NoSePudo Football Players API
  description: Public player catalog and ingestion API for top European football leagues.
  version: 1.0.0
paths:
  /players:
    get:
      summary: List players with pagination and filters
      description: |
        Retrieves a paginated list of football players.
        Per system requirements and Principle V, the list view returns summary DTOs
        containing only the name, current team (club), league, and position.
      operationId: getPlayers
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of players per page
        - name: league
          in: query
          required: false
          schema:
            type: string
          description: Filter by league name or competition code (e.g., PL, Premier League)
        - name: club
          in: query
          required: false
          schema:
            type: string
          description: Filter by club / team name
        - name: position
          in: query
          required: false
          schema:
            type: string
          description: Filter by playing position (e.g., Goalkeeper, Defender, Midfielder, Attacker)
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Free-text search on player name
        - name: includeInactive
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Set to true to include departed/inactive players in historical catalog view
      responses:
        '200':
          description: Paginated player summary listing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPlayerListResponse'
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /players/{id}:
    get:
      summary: Get player profile details
      description: |
        Retrieves complete attributes for a single player by their internal ID.
        Per system requirements and Principle V, returns the full PlayerDetailResponse DTO.
      operationId: getPlayerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Internal player database ID
      responses:
        '200':
          description: Full player details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerDetailResponse'
        '400':
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /players/sync:
    post:
      summary: Trigger player synchronization from football-data.org
      description: |
        Triggers ingestion of active players across the 5 top European leagues.
        Retrieves squads from football-data.org, updates the local database,
        soft-deactivates departed players, and records transactional audit logs.
      operationId: syncPlayers
      responses:
        '200':
          description: Synchronization finished successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncPlayersResponse'
        '429':
          description: External provider rate limit encountered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Synchronization failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    PlayerListItemResponse:
      type: object
      required:
        - id
        - name
        - club
        - league
        - position
      properties:
        id:
          type: integer
          format: int64
          example: 101
        name:
          type: string
          example: "Bukayo Saka"
        club:
          type: string
          example: "Arsenal FC"
        league:
          type: string
          example: "Premier League"
        position:
          type: string
          example: "Attacker"

    PaginatedPlayerListResponse:
      type: object
      required:
        - items
        - page
        - limit
        - total
        - totalPages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PlayerListItemResponse'
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          format: int64
          example: 184
        totalPages:
          type: integer
          example: 10

    PlayerDetailResponse:
      type: object
      required:
        - id
        - externalId
        - name
        - club
        - league
        - leagueCode
        - position
        - active
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          example: 101
        externalId:
          type: integer
          format: int64
          example: 7821
        name:
          type: string
          example: "Bukayo Saka"
        club:
          type: string
          example: "Arsenal FC"
        league:
          type: string
          example: "Premier League"
        leagueCode:
          type: string
          example: "PL"
        position:
          type: string
          example: "Attacker"
        dateOfBirth:
          type: string
          nullable: true
          example: "2001-09-05"
        nationality:
          type: string
          nullable: true
          example: "England"
        shirtNumber:
          type: integer
          nullable: true
          example: 7
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2026-09-28T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-09-28T12:30:00Z"

    SyncPlayersResponse:
      type: object
      required:
        - status
        - message
        - totalProcessed
        - totalUpdated
        - totalDeactivated
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Player synchronization completed successfully"
        totalProcessed:
          type: integer
          example: 2500
        totalUpdated:
          type: integer
          example: 45
        totalDeactivated:
          type: integer
          example: 12

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid query parameter 'page'"